refactor: BaseEntity 패키지 수정
